Add support for setting speeds via serial

+++ src/lib_wrappers/Distance.hpp
@@ -3,9 +3,11 @@
 
 #include <stdint.h>
 
-namespace Libs {
+namespace Libs
+{
 
-class Distance {
+class Distance
+{
 public:
   constexpr Distance() : m_micrometers(0) {}
 
@@ -17,33 +19,40 @@ public:
       : m_micrometers((meters * METER_TO_MICRO) +
                       (millimeters * MILLI_TO_MICRO)) {}
 
-  constexpr bool operator==(const Distance &other) const {
+  constexpr bool operator==(const Distance &other) const
+  {
     return m_micrometers == other.m_micrometers;
   }
 
-  constexpr Distance operator+(const Distance &other) const {
+  constexpr Distance operator+(const Distance &other) const
+  {
     return Distance(m_micrometers + other.m_micrometers);
   }
 
-  constexpr Distance operator-(const Distance &other) const {
+  constexpr Distance operator-(const Distance &other) const
+  {
     return Distance(m_micrometers - other.m_micrometers);
   }
 
-  constexpr Distance operator*(const int32_t multi) const {
+  constexpr Distance operator*(const int32_t multi) const
+  {
     return Distance(m_micrometers * multi);
   }
 
-  constexpr Distance operator/(const int32_t div) const {
+  constexpr Distance operator/(const int32_t div) const
+  {
     return Distance(m_micrometers / div);
   }
 
   constexpr int32_t micrometers() const { return m_micrometers; }
 
-  constexpr int32_t millimeters() const {
+  constexpr int32_t millimeters() const
+  {
     return m_micrometers / MILLI_TO_MICRO;
   }
-  constexpr double meters() const {
-    return m_micrometers / (double)METER_TO_MICRO;
+  constexpr float meters() const
+  {
+    return m_micrometers / (float)METER_TO_MICRO;
   }
 
   constexpr int32_t native() const { return m_micrometers; }
@@ -56,11 +65,13 @@ private:
   int32_t m_micrometers{0};
 }; // namespace Libs
 
-constexpr Distance operator"" _m(unsigned long long meters) {
+constexpr Distance operator"" _m(unsigned long long meters)
+{
   return Distance(meters, 0);
 }
 
-constexpr Distance operator"" _m(long double meters) {
+constexpr Distance operator"" _m(long double meters)
+{
   // Has to be a long double as per the standard
   return Distance(meters, 0);
 }
